+++ src/components/SharedComponents/helpers/downloadAndShareHelpers/downloadLessonResources.tsx
@@ -8,14 +8,14 @@ const downloadLessonResources = async ({
   selectedResourceTypes,
   selectedAdditionalFilesIds,
   isLegacyDownload,
-  authFlagEnabled,
+  authRequired,
   authToken,
 }: {
   lessonSlug: string;
   selectedResourceTypes: (DownloadResourceType | "worksheet-pdf-questions")[]; // FIXME: a new solution is required for types which are shared across different journeys . Also should the downloads schemas be added to oak-curriculum schema?
   selectedAdditionalFilesIds?: number[];
   isLegacyDownload: boolean;
-  authFlagEnabled?: boolean;
+  authRequired?: boolean;
   authToken?: string | null;
 }) => {
   if (selectedResourceTypes?.length === 0) {
@@ -31,7 +31,7 @@ const downloadLessonResources = async ({
     selection,
     additionalFilesIdsSelection,
     isLegacyDownload,
-    authFlagEnabled,
+    authRequired,
     authToken,
   });