+++ src/components/TeacherComponents/SavingSignedOutModal/SavingSignedOutModal.tsx
@@ -37,6 +37,9 @@ const SavingSignedOutModalContent = () => {
         quickly sign up and start saving.
       </OakP>
       <LoginRequiredButton
+        loginRequired={true}
+        geoRestricted={false}
+        isBehindFeatureFlag={false}
         onboardingProps={{ name: "Finish sign up" }}
         width="100%"
       />