+++ src/pages/teachers/lessons/[lessonSlug]/downloads.tsx
@@ -44,6 +44,7 @@ const LessonDownloadsCanonicalPage = (
           description: "Lesson downloads",
         }),
         noIndex: true,
+        noFollow: true,
       }}
     >
       <LessonDownloads isCanonical lesson={curriculumData} />